# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Site API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.local",
            "description": "Local HTTPS server"
        }
    ],
    "paths": {
...
        "/api/telegram": {
            "get": {
                "tags": [
                    "Telegram"
                ],
                "summary": "Send Telegram message via signed query params",
                "description": "Legacy-compatible endpoint. Use msg/domain/s parameters to send outgoing messages.",
                "operationId": "apiTelegramGet",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "msg",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "domain",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "s",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Telegram chat id, can be repeated as array id[]=1&id[]=2",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "answer",
                        "in": "query",
                        "description": "Set to json to force JSON response",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "json"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Command handled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "res": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "text": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation-like failure, for example missing token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "res": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "text": {
                                            "type": "string",
                                            "example": "token is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Telegram"
                ],
                "summary": "Telegram webhook handler and sender",
                "description": "Accepts webhook updates from Telegram and also supports outgoing message mode with legacy payload fields.",
                "operationId": "apiTelegramPost",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "msg": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "domain": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "s": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "id": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "nullable": true
                                    },
                                    "answer": {
                                        "type": "string",
                                        "example": "json",
                                        "nullable": true
                                    },
                                    "message": {
                                        "properties": {
                                            "text": {
                                                "type": "string",
                                                "example": "/start",
                                                "nullable": true
                                            },
                                            "from": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 123456789
                                                    },
                                                    "username": {
                                                        "type": "string",
                                                        "example": "username",
                                                        "nullable": true
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "Ivan",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook handled or message processed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "res": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "text": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Missing required token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "res": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "text": {
                                            "type": "string",
                                            "example": "token is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}